refactor: move parsePhoneNo function to common
